Sat, Nov 4, 2023, 12:16 PM -05:00

+++ easter.py
@@ -1,6 +1,6 @@
 
 print ('Enter a year:')
-y = int(input())
+y = float(input())
 
 
 a = y / 19
